Add EditCommentView as a sheet
Didn't really want to do that, but didn't really have a choice. When we try to submit the changes via the CommentView, it will cause changes to the viewmodel which will re-render the view. Concurrently, we need to refresh the state of the view (to clear the edit fields). These two concurrent changes can cause unexpected behaviour to the view (based on the warnings). Hence, I used the sheet solution as per our other views, though it is a bit unsightly.