create ViewModelFactory
Also renamed the vms in different views to viewModel, for consistency. Technically each view should just have its own VM.
Shifted some logic from views into VMs.

Wherever a new ViewModel is to be created, replaced with the Factory's methods.
The methods should either use an existing object instance in the view for rendering, an existing viewmodel or both to create a new VM.
Purpose is so the Views work with what they have or from the Factory.